Make qreal values explicit

+++ tests/auto/qgeocoordinate/tst_qgeocoordinate.cpp
@@ -456,15 +456,15 @@ private slots:
         QTest::addColumn<qreal>("distance");
 
         QTest::newRow("invalid coord 1")
-                << QGeoCoordinate() << BRISBANE << 0.0;
+                << QGeoCoordinate() << BRISBANE << qreal(0.0);
         QTest::newRow("invalid coord 2")
-                << BRISBANE << QGeoCoordinate() << 0.0;
+                << BRISBANE << QGeoCoordinate() << qreal(0.0);
         QTest::newRow("brisbane -> melbourne")
-                << BRISBANE << MELBOURNE << 1374820.162;
+                << BRISBANE << MELBOURNE << qreal(1374820.162);
         QTest::newRow("london -> new york")
-                << LONDON << NEW_YORK << 5570538.499;
+                << LONDON << NEW_YORK << qreal(5570538.499);
         QTest::newRow("north pole -> south pole")
-                << NORTH_POLE << SOUTH_POLE << 20015109.42;
+                << NORTH_POLE << SOUTH_POLE << qreal(20015109.42);
     }
 
     void azimuthTo()
@@ -483,15 +483,15 @@ private slots:
         QTest::addColumn<qreal>("azimuth");
 
         QTest::newRow("invalid coord 1")
-                << QGeoCoordinate() << BRISBANE << 0.0;
+                << QGeoCoordinate() << BRISBANE << qreal(0.0);
         QTest::newRow("invalid coord 2")
-                << BRISBANE << QGeoCoordinate() << 0.0;
+                << BRISBANE << QGeoCoordinate() << qreal(0.0);
         QTest::newRow("brisbane -> melbourne")
-                << BRISBANE << MELBOURNE << 211.1717;
+                << BRISBANE << MELBOURNE << qreal(211.1717);
         QTest::newRow("london -> new york")
-                << LONDON << NEW_YORK << 288.3389;
+                << LONDON << NEW_YORK << qreal(288.3389);
         QTest::newRow("north pole -> south pole")
-                << NORTH_POLE << SOUTH_POLE << 180.0;
+                << NORTH_POLE << SOUTH_POLE << qreal(180.0);
     }
 
     void degreesToString()
